.archive/noodle-poc-fase1/pyproject.toml
[project]
name = "noodle-poc-fase1"
version = "0.1.0"
description = "NoodleCore Distributed Inference - Fase 1: Observability Engine"
authors = [
    {name = "NoodleCore Team", email = "team@noodlecore.dev"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "psutil>=5.9.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0"
]
dashboard = [
    "dash>=2.10.0",
    "dash-bootstrap-components>=1.4.0",
    "plotly>=5.14.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
